Upgrade to using a proper MD5 module extension
Calling an external utility is sometimes a little faster, but in this
case it’s not worth the pain when you can’t control or properly escape
file names’ weird punctuation. In this case, the module is compiled XS
working on binary file handles without overhead of spawning external
process, so no degradation of speed anyway.